Split out factory task function to fix imports.
