+++ futaba/cogs/navi/core.py
@@ -26,8 +26,8 @@ from discord.ext import commands
 from futaba import permissions
 from futaba.exceptions import CommandFailed
 from futaba.utils import escape_backticks, fancy_timedelta
-from .task import SendMessageTask, build_navi_task
 from .scheduler import NaviScheduler
+from .task import SendMessageTask, build_navi_task
 
 logger = logging.getLogger(__name__)
